refactor: make ParseDicomDirResult as class for better documentation (#394)
I wanted to add a docstring to `ParseDicomDirResult` but because it was
a named tuple and not a class, I had to change it.

Now `ParseDicomDirResult` is a subclass of the named tuple class. It is
functionally the exact same except now I can document it.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **Documentation**
- Improved documentation for output details, providing clearer
descriptions of returned data fields. No changes to user-facing features
or data formats.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Declan Michael Korda <DKORDA@uwo.ca>
Co-authored-by: Jermiah Joseph <jermiahjoseph98@gmail.com>
Co-authored-by: Jermiah Joseph <44614774+jjjermiah@users.noreply.github.com>